cypress cleanup
