+++ package.json
@@ -29,7 +29,6 @@
     "@typescript-eslint/eslint-plugin": "5.18.0",
     "@typescript-eslint/parser": "5.18.0",
     "aws-sdk": "^2.1310.0",
-    "cypress-tests": "file:tests/cypress",
     "dotenv": "^8.2.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
