+++ tests/cypress/cypress/e2e/a11y/LBWCHpage.spec.ts
@@ -1,11 +1,13 @@
 import { testingYear } from "../../../support/constants";
 
-describe.skip("LBW-CH Page 508 Compliance Test", () => {
+describe("LBW-CH Page 508 Compliance Test", () => {
   it("Check a11y on LBW-CH Page", () => {
     cy.login();
     cy.selectYear(testingYear);
     cy.goToChildCoreSetMeasures();
-    cy.goToMeasure("LBW-CH");
+    cy.get(`[data-cy="LBW-CH"]`, { timeout: 10000 })
+      .should("be.visible")
+      .click();
     cy.checkA11yOfPage();
   });
 });
