+++ tests/cypress/cypress/e2e/a11y/LRCDCH.spec.ts
@@ -1,11 +1,13 @@
 import { testingYear } from "../../../support/constants";
 
-describe.skip("LRCD-CH Page 508 Compliance Test", () => {
+describe("LRCD-CH Page 508 Compliance Test", () => {
   it("Check a11y on LRCD-CH Page", () => {
     cy.login();
     cy.selectYear(testingYear);
     cy.goToChildCoreSetMeasures();
-    cy.goToMeasure("LRCD-CH");
+    cy.get(`[data-cy="LRCD-CH"]`, { timeout: 10000 })
+      .should("be.visible")
+      .click();
     cy.checkA11yOfPage();
   });
 });
