+++ tests/cypress/cypress/e2e/a11y/nciddsPage.spec.ts
@@ -1,11 +1,13 @@
 import { testingYear } from "../../../support/constants";
 
-describe.skip("NCIDDS-AD Page 508 Compliance Test", () => {
+describe("NCIDDS-AD Page 508 Compliance Test", () => {
   it("Check a11y on NCIDDSAD Page", () => {
     cy.login();
     cy.selectYear(testingYear);
     cy.goToAdultMeasures();
-    cy.goToMeasure("NCIDDS-AD");
+    cy.get(`[data-cy="NCIDDS-AD"]`, { timeout: 10000 })
+      .should("be.visible")
+      .click();
     cy.checkA11yOfPage();
   });
 });
