+++ tests/cypress/cypress/e2e/features/export_all_measures.spec.ts
@@ -1,5 +1,4 @@
-import { testingYear } from "../../../support/constants";
-import { measureAbbrList2023 } from "../../../support/commands/commands";
+import { measureAbbrList2023, testingYear } from "../../../support/constants";
 
 describe("Export All Measures", () => {
   beforeEach(() => {
