+++ tests/cypress/support/commands.ts
@@ -314,85 +314,3 @@ Cypress.Commands.add("SSHHdisplaysCorrectSections", () => {
     '[data-cy="Additional Notes/Comments on the measure (optional)"]'
   ).should("be.visible");
 });
-
-export const measureAbbrList2023 = {
-  ADULT: [
-    "AAB-AD",
-    "AMM-AD",
-    "AMR-AD",
-    "BCS-AD",
-    "CBP-AD",
-    "CCP-AD",
-    "CCS-AD",
-    "CCW-AD",
-    "CDF-AD",
-    "CHL-AD",
-    "COB-AD",
-    "COL-AD",
-    "CPA-AD",
-    "FUA-AD",
-    "FUH-AD",
-    "FUM-AD",
-    "FVA-AD",
-    "HBD-AD",
-    "HPCMI-AD",
-    "HVL-AD",
-    "IET-AD",
-    "MSC-AD",
-    "NCIDDS-AD", // complete on creation
-    "OHD-AD",
-    "OUD-AD",
-    "PCR-AD",
-    "PPC-AD",
-    "PQI01-AD",
-    "PQI05-AD",
-    "PQI08-AD",
-    "PQI15-AD",
-    "SAA-AD",
-    "SSD-AD",
-  ],
-  CHILD: [
-    "AAB-CH",
-    "ADD-CH",
-    "AMB-CH",
-    "AMR-CH",
-    "APM-CH",
-    "APP-CH",
-    "CCP-CH",
-    "CCW-CH",
-    "CDF-CH",
-    "CHL-CH",
-    "CIS-CH",
-    "CPC-CH",
-    "DEV-CH",
-    "FUA-CH",
-    "FUH-CH",
-    "FUM-CH",
-    "IMA-CH",
-    "LBW-CH", // complete on creation
-    "LRCD-CH", // complete on creation
-    "LSC-CH",
-    "OEV-CH",
-    "PPC-CH",
-    "SFM-CH",
-    "TFL-CH",
-    "W30-CH",
-    "WCC-CH",
-    "WCV-CH",
-  ],
-  HEALTH_HOME: [
-    "AIF-HH",
-    "AMB-HH",
-    "CBP-HH",
-    "CDF-HH",
-    "COL-HH",
-    "FUA-HH",
-    "FUH-HH",
-    "FUM-HH",
-    "IET-HH",
-    "IU-HH",
-    "OUD-HH",
-    "PCR-HH",
-    "PQI92-HH",
-  ],
-};
