--- tests/cypress/support/commands/index.js
@@ -1,2 +1,0 @@
-import "./commands";
-import "./performanceMeasure";
