--- tests/cypress/support/commands/performanceMeasure/index.ts
@@ -1,8 +1,0 @@
-Cypress.Commands.add(
-  "fillInAllPerformanceMeasureRates",
-  ({ qualifiers, categories, defaultAllValue = 2, values, incrementing }) => {
-    const fillArray = values ?? [
-      [defaultAllValue, defaultAllValue, defaultAllValue],
-    ];
-  }
-);
