--- tests/cypress/support/commands/performanceMeasure/performanceMeasure.d.ts
@@ -1,12 +1,0 @@
-declare namespace Cypress {
-  interface Chainable {
-    /** fills all performance measure NDR sets with optional values */
-    fillInAllPerformanceMeasureRates(props: {
-      qualifiers: string[];
-      categories: string[];
-      defaultAllValue?: number;
-      values?: number[][];
-      incrementing?: boolean;
-    }): Chainable<Element>;
-  }
-}
